+++ lib/src/activsy.dart
@@ -31,10 +31,15 @@ class Activsy {
     return _instance;
   }
 
-  /// start and set the timer [startTimer]
+  /// init timer
+  void init() {
+    start();
+  }
+
+  /// start and set the timer [start]
   /// this function to be called after the [config] method
   /// otherwise you will throw an exception
-  static void startTimer() {
+  static void start() {
     if (_instance._showLog) log('activsy::start', time: DateTime.now());
 
     if (_instance._callback == null)
@@ -52,8 +57,8 @@ class Activsy {
     });
   }
 
-  /// stop the timer [stopTimer]
-  static void stopTimer() {
+  /// stop the timer [cancel]
+  static void cancel() {
     if (_instance._showLog) log('activsy::stop', time: DateTime.now());
 
     var timer = _instance._timer;
@@ -75,9 +80,9 @@ class Activsy {
 
     _instance._seconds = _seconds;
 
-    stopTimer();
+    cancel();
 
-    startTimer();
+    start();
   }
 
   /// There are several reasons to trigger the noActivity method before the stipulated timer [trigger]
@@ -87,7 +92,7 @@ class Activsy {
     if (_instance._callback == null)
       throw Exception('this function to be called after the config method');
 
-    stopTimer();
+    cancel();
 
     _instance._callback!();
   }
@@ -96,4 +101,8 @@ class Activsy {
   /// this function to be called after the [config] method
   /// otherwise you will throw an exception
   static bool get isInitialized => _instance._callback != null;
+
+  /// checks monitoring is active [isActive]
+  static bool get isActive =>
+      (_instance._timer != null && _instance._timer!.isActive);
 }