+++ lib/src/activsy_widget.dart
@@ -5,11 +5,24 @@ class ActivsyWidget extends StatelessWidget {
   /// [builder]  principal widget
   final Widget Function(BuildContext) builder;
 
+  final Function(PointerEvent)? onEvent;
+
   /// [detectedMouseAction] enable/disable mouse detector
   final bool detectedMouseAction;
 
-  ActivsyWidget({required this.builder, this.detectedMouseAction = false})
-      : assert(Activsy.isInitialized, 'activsy no initialized');
+  ActivsyWidget(
+      {required this.builder, this.detectedMouseAction = false, this.onEvent})
+      : assert(Activsy.isInitialized, () {
+          throw FlutterError(
+              'activsy no initialized.\nMust call the Activsy.config() before the ActivsyWidget');
+          return true;
+        }());
+
+  /// notify when has event
+  void _onEvent(PointerEvent event) {
+    Activsy.reset();
+    if (onEvent != null && Activsy.isActive) onEvent!(event);
+  }
 
   /// sets up the widget to be shown [_builder]
   Widget _builder(BuildContext context) {
@@ -17,9 +30,9 @@ class ActivsyWidget extends StatelessWidget {
     if (detectedMouseAction) {
       return MouseRegion(
         child: builder(context),
-        onEnter: (_) => Activsy.reset(),
-        onHover: (_) => Activsy.reset(),
-        onExit: (_) => Activsy.reset(),
+        onEnter: _onEvent,
+        onHover: _onEvent,
+        onExit: _onEvent,
       );
     }
     return builder(context);
@@ -28,13 +41,15 @@ class ActivsyWidget extends StatelessWidget {
   /// Here we hear all the events of gestures in our application
   /// Whenever any gesture event is detected the wait time and restarted
   @override
-  Widget build(BuildContext context) => Listener(
-        child: _builder(context),
-        onPointerDown: (_) => Activsy.reset(),
-        onPointerUp: (_) => Activsy.reset(),
-        onPointerSignal: (_) => Activsy.reset(),
-        onPointerMove: (_) => Activsy.reset(),
-        onPointerCancel: (_) => Activsy.reset(),
-        onPointerHover: (_) => Activsy.reset(),
-      );
+  Widget build(BuildContext context) {
+    return Listener(
+      child: _builder(context),
+      onPointerDown: _onEvent,
+      onPointerUp: _onEvent,
+      onPointerSignal: _onEvent,
+      onPointerMove: _onEvent,
+      onPointerCancel: _onEvent,
+      onPointerHover: _onEvent,
+    );
+  }
 }