+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 * Stop timer
 * Reset timer 
 * Change wait time
+
+## 1.0.1
+* Intercept events
+* Change in the name of methods
+* Check active monitoring